finish meal plan data handler, view handler, adapter, RecipeSearchActivity (#99)
* finish meal plan data handler, view handler, adapter (but no data for testing now); add RecipeSearchActivity and meal plan can use this to go there

* using setProperties to modify meal plan recipe list

+++ app/src/main/java/com/github/kuma/grocerymanager/MealplanDataHandler.java
@@ -1,6 +1,8 @@
 package com.github.kuma.grocerymanager;
 
 import android.content.Context;
+import android.util.Pair;
+
 import com.github.kuma.data.db.DbDocument;
 import com.github.kuma.data.db.SimpleDbInterface;
 import com.github.kuma.db_object.Mealplan;
@@ -14,17 +16,71 @@ public class MealplanDataHandler
 {
     private Context context;
     private List<DbDocument> data;
-    private HashMap<Integer,MealplanRecipe> viewData;
-    private HashMap<Integer,Mealplan> javaData;
+    private HashMap<Pair<Integer,Integer>,List<MealplanRecipe>> viewData;
+    private HashMap<Pair<Integer,Integer>, String> map_of_mealplanId;
     private Date today;
 
-    public MealplanDataHandler(Context context) throws Exception
+    public MealplanDataHandler(Context context)
+    {
+        this.context = context;
+        viewData = new HashMap<Pair<Integer, Integer>, List<MealplanRecipe>>();
+        map_of_mealplanId = new HashMap<Pair<Integer, Integer>, String>();
+        today = new Date();
+    }
+
+    public void deleteRecipe (int day, int period, String name) throws Exception
+    {
+        String mealplanId = map_of_mealplanId.get(new Pair<Integer, Integer>(day,period));
+        DbDocument mealplan_dbdoc = new DbDocument(context,mealplanId);
+
+        // check matched recipe
+        List<String> recipeIds = (List<String>)mealplan_dbdoc.getProperty("recipesIds");
+        for(String id : recipeIds)
+        {
+            DbDocument recipe_dbdoc = new DbDocument(context,id);
+            String recipeName = (String)recipe_dbdoc.getProperty("name");
+            if(name.equals(recipeName))
+            {
+                recipeIds.remove(id);
+                break;
+            }
+        }
+        // check matched food
+        List<String> foods = (List<String>) mealplan_dbdoc.getProperty("foodNames");
+        for(String foodName : foods)
+        {
+            if(name.equals(foodName))
+            {
+                foods.remove(foodName);
+                break;
+            }
+        }
+
+        // update meal plan data
+        HashMap<String,Object> newProperties = new HashMap<String, Object>();
+        newProperties.put("recipeIds",recipeIds);
+        newProperties.put("foodNames",foods);
+        mealplan_dbdoc.setProperties(newProperties);
+
+    }
+
+    public void loadData() throws  Exception
     {
         data = SimpleDbInterface.getAllMealPlanDocuments(context);
-        viewData = new HashMap<Integer, MealplanRecipe>();
-        today = new Date();
+    }
 
-        validateData(); // remove any meal plan data before today
+    public HashMap<Pair<Integer,Integer>,List<MealplanRecipe>> getViewData() throws Exception
+    {
+        validateData(); // remove any meal plan data before today and generate view data, map of mealplan id
+        return viewData;
+    }
+
+    private int getMealTypeNum (String mealType)
+    {
+        if(mealType.equals(Mealplan.BREAKFAST)) return 1;
+        else if(mealType.equals(Mealplan.LUNCH)) return 2;
+        else return 3;
+
     }
 
     private void validateData() throws Exception
@@ -35,22 +91,61 @@ public class MealplanDataHandler
             {
                 Date planDate = (Date) dbDoc.getProperty("date");
                 int daysBetween = DateUtils.getDaysBetween(today,planDate);
+                String mealType = (String)dbDoc.getProperty("mealType");
+                int mealTypeNum = getMealTypeNum(mealType);
+                String mealplanId = (String) dbDoc.getProperty("id");
 
-                if(daysBetween < 0)
+                Pair<Integer,Integer> mapkey = new Pair<Integer, Integer>(daysBetween,mealTypeNum);
+                map_of_mealplanId.put(mapkey,mealplanId);
+
+                if(daysBetween < 0)     // when mealplan expires more than two weeks (14 days,default), remove the meal plan entry
+                                        // days can also be modified in setting
                 {
-                    dbDoc.delete();
+                    if(daysBetween <= -14)
+                    {
+                        dbDoc.delete();
+                    }
                 }
                 else{
-                    // construct javaData and viewData
-                    List<String> list_of_recipe_names = (List<String>) dbDoc.getProperty("foodNames");
+                    // construct viewData
+                    List<String> list_of_recipe_ids = (List<String>) dbDoc.getProperty("recipesIds");
+                    List<String> list_of_foodNames = (List<String>) dbDoc.getProperty("foodNames");
 
-                    for(String recipe : list_of_recipe_names)
+                    List<MealplanRecipe> list_of_recipe = viewData.get(mapkey);
+
+                    if(list_of_recipe == null)
                     {
-
+                        list_of_recipe = new ArrayList<MealplanRecipe>();
                     }
+                    // add recipes
+                    for(String recipeId : list_of_recipe_ids)
+                    {
+                        DbDocument recipeDbDoc = new DbDocument(context,recipeId);
+                        String mealName = (String)recipeDbDoc.getProperty("name");
 
+                        MealplanRecipe meal = new MealplanRecipe(daysBetween,mealTypeNum,mealName);
+
+                        if(list_of_recipe == null)
+                        {
+                            list_of_recipe = new ArrayList<MealplanRecipe>();
+                        }
+
+                        list_of_recipe.add(meal);
+                        viewData.put(mapkey,list_of_recipe);
+                    }
+                    // add foods (i.e. apples, eggs)
+                    for(String food : list_of_foodNames)
+                    {
+                        if(list_of_recipe == null)
+                        {
+                            list_of_recipe = new ArrayList<MealplanRecipe>();
+                        }
+                        MealplanRecipe meal = new MealplanRecipe(daysBetween,mealTypeNum,food);
+
+                        list_of_recipe.add(meal);
+                        viewData.put(mapkey,list_of_recipe);
+                    }
                 }
-
             }
         }
     }
